+++ .config/nvim/lua/vim-options.lua
@@ -57,7 +57,7 @@ vim.keymap.set("n", "<leader>yp", yank_path("%:."), { desc = "Yank relative path
 vim.keymap.set("n", "<leader>yP", yank_path("%:p"), { desc = "Yank absolute path" })
 
 -- Split navigation keymaps
-vim.keymap.set("n", "<C-h>", "<C-w>h", {})
-vim.keymap.set("n", "<C-j>", "<C-w>j", {})
-vim.keymap.set("n", "<C-k>", "<C-w>k", {})
-vim.keymap.set("n", "<C-l>", "<C-w>l", {})
+vim.keymap.set("n", "<C-h>", "<C-w>h", { desc = "Window left" })
+vim.keymap.set("n", "<C-j>", "<C-w>j", { desc = "Window down" })
+vim.keymap.set("n", "<C-k>", "<C-w>k", { desc = "Window up" })
+vim.keymap.set("n", "<C-l>", "<C-w>l", { desc = "Window right" })